Further polish the base class implementation.
