+++ api/tests_v2/common_import.py
@@ -31,7 +31,7 @@ except ImportError:
 package_path = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(package_path)
 
-from common.paddle_api_benchmark import PaddleAPIBenchmarkBase
+from common.paddle_op_benchmark import PaddleAPIBenchmarkBase
 from common.tensorflow_api_benchmark import TensorflowAPIBenchmarkBase
 from common.api_param import APIConfig
 from common.main import test_main, test_main_without_json
